+++ Inferno/Source/Graphics/Font/Font.hpp
@@ -14,6 +14,6 @@ void SetGlobalFont(PSFFont* temp);
 void putc(char c, unsigned int xOff, unsigned int yOff);
 void printf(const char* string, ...);
 void printf(long long int i);
-void hprintf(unsigned long long int i);
+void hprintf(unsigned int i);
 
 extern PSFFont* font;
